+++ src/Email/Templates/download-expiring.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'subject' => 'Your Download Link Is Expiring Soon',
+    'body' => "Hi {{client_name}},\n\n"
+        . "Just a quick reminder that your download link will expire soon.\n\n"
+        . "If you need more time, you can request a new link using the client portal.\n\n"
+        . "Best,\n{{studio_name}}",
+];